apply MIN_BOOST because otherwise MAP gives stupid results
`ankiFitEasyHardMAP([4, 3, 1], [10., 13., 19.], **kws)` would return
hl=25 hours with b=0.5 because that maximized posterior.

This is because an Exponential rv's mode is at 0, so MAP was going to
the min boost allowed by SHGO.

I didn't want to set the alpha parameter of boost's Gamma rv (recall
that an Exponential is a Gamma with fixed alpha=1) to make the mode of
boost at 1.0 because then monotonicity of easy/normal/hard broke.

Also removed a lot of legacy parameters like clamp (always clamp) and
Gamma-vs-Expon